Bugfix for completion code. Also some minor code changes.

+++ CxImage/ximaico.cpp
@@ -322,7 +322,7 @@ bool CxImageICO::Encode(CxFile * hFile, bool bAppend, int32_t nPageCount)
 	int32_t nPages = nPageCount;
 	if (nPages<1) nPages = 1;
 
-	ICONHEADER icon_header={0,1,nPages};
+	ICONHEADER icon_header = {0, 1, (uint16_t)nPages};
 
 	if (!bAppend)
 		m_dwImageOffset = sizeof(ICONHEADER) + nPages * sizeof(ICONDIRENTRY);
@@ -344,8 +344,8 @@ bool CxImageICO::Encode(CxFile * hFile, bool bAppend, int32_t nPageCount)
 		head.biWidth,
 		2*head.biHeight,
 		1,
-		(uint16_t)bitcount,
-		0, imagesize,
+		(WORD)bitcount,
+		0, (DWORD)imagesize,
 		0, 0, 0, 0
 	};
